Refactor UnixFileSystem
