+++ Source/Engine/Platform/Apple/AppleFileSystem.cpp
@@ -7,10 +7,6 @@
 #include "Engine/Platform/File.h"
 #include "Engine/Core/Types/String.h"
 #include "Engine/Core/Types/StringView.h"
-#include "Engine/Core/Types/TimeSpan.h"
-#include "Engine/Core/Collections/Array.h"
-#include "Engine/Core/Math/Math.h"
-#include "Engine/Core/Log.h"
 #include "Engine/Utilities/StringConverter.h"
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -18,280 +14,8 @@
 #include <stdio.h>
 #include <cerrno>
 #include <dirent.h>
-#include <string.h>
 #include <stdlib.h>
 #include <fcntl.h>
-
-const DateTime UnixEpoch(1970, 1, 1);
-
-bool AppleFileSystem::CreateDirectory(const StringView& path)
-{
-    const StringAsANSI<> pathAnsi(*path, path.Length());
-
-    // Skip if already exists
-    struct stat fileInfo;
-    if (stat(pathAnsi.Get(), &fileInfo) != -1 && S_ISDIR(fileInfo.st_mode))
-    {
-        return false;
-    }
-
-    // Recursively do it all again for the parent directory, if any
-    const int32 slashIndex = path.FindLast('/');
-    if (slashIndex > 1)
-    {
-        if (CreateDirectory(path.Substring(0, slashIndex)))
-        {
-            return true;
-        }
-    }
-
-    // Create the last directory on the path (the recursive calls will have taken care of the parent directories by now)
-    return mkdir(pathAnsi.Get(), 0755) != 0 && errno != EEXIST;
-}
-
-bool DeletePathTree(const char* path)
-{
-    size_t pathLength;
-    DIR* dir;
-    struct stat statPath, statEntry;
-    struct dirent* entry;
-
-    // Stat for the path
-    stat(path, &statPath);
-
-    // If path does not exists or is not dir - exit with status -1
-    if (S_ISDIR(statPath.st_mode) == 0)
-    {
-        // Is not directory
-        return true;
-    }
-
-    // If not possible to read the directory for this user
-    if ((dir = opendir(path)) == NULL)
-    {
-        // Cannot open directory
-        return true;
-    }
-
-    // The length of the path
-    pathLength = strlen(path);
-
-    // Iteration through entries in the directory
-    while ((entry = readdir(dir)) != NULL)
-    {
-        // Skip entries "." and ".."
-        if (!strcmp(entry->d_name, ".") || !strcmp(entry->d_name, ".."))
-            continue;
-
-        // Determinate a full path of an entry
-        char full_path[256];
-        ASSERT(pathLength + strlen(entry->d_name) < ARRAY_COUNT(full_path));
-        strcpy(full_path, path);
-        strcat(full_path, "/");
-        strcat(full_path, entry->d_name);
-
-        // Stat for the entry
-        stat(full_path, &statEntry);
-
-        // Recursively remove a nested directory
-        if (S_ISDIR(statEntry.st_mode) != 0)
-        {
-            if (DeletePathTree(full_path))
-                return true;
-            continue;
-        }
-
-        // Remove a file object
-        if (unlink(full_path) != 0)
-            return true;
-    }
-
-    // Remove the devastated directory and close the object of it
-    if (rmdir(path) != 0)
-        return true;
-
-    closedir(dir);
-
-    return false;
-}
-
-bool AppleFileSystem::DeleteDirectory(const String& path, bool deleteContents)
-{
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (deleteContents)
-        return DeletePathTree(pathANSI.Get());
-    return rmdir(pathANSI.Get()) != 0;
-}
-
-bool AppleFileSystem::DirectoryExists(const StringView& path)
-{
-    struct stat fileInfo;
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (stat(pathANSI.Get(), &fileInfo) != -1)
-    {
-        return S_ISDIR(fileInfo.st_mode);
-    }
-    return false;
-}
-
-bool AppleFileSystem::DirectoryGetFiles(Array<String>& results, const String& path, const Char* searchPattern, DirectorySearchOption option)
-{
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    const StringAsANSI<> searchPatternANSI(searchPattern);
-
-    // Check if use only top directory
-    if (option == DirectorySearchOption::TopDirectoryOnly)
-        return getFilesFromDirectoryTop(results, pathANSI.Get(), searchPatternANSI.Get());
-    return getFilesFromDirectoryAll(results, pathANSI.Get(), searchPatternANSI.Get());
-}
-
-bool AppleFileSystem::GetChildDirectories(Array<String>& results, const String& path)
-{
-    size_t pathLength;
-    DIR* dir;
-    struct stat statPath, statEntry;
-    struct dirent* entry;
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    const char* pathStr = pathANSI.Get();
-
-    // Stat for the path
-    stat(pathStr, &statPath);
-
-    // If path does not exist or is not dir - exit with status -1
-    if (S_ISDIR(statPath.st_mode) == 0)
-    {
-        // Is not directory
-        return true;
-    }
-
-    // If not possible to read the directory for this user
-    if ((dir = opendir(pathStr)) == NULL)
-    {
-        // Cannot open directory
-        return true;
-    }
-
-    // The length of the path
-    pathLength = strlen(pathStr);
-
-    // Iteration through entries in the directory
-    while ((entry = readdir(dir)) != NULL)
-    {
-        // Skip entries "." and ".."
-        if (!strcmp(entry->d_name, ".") || !strcmp(entry->d_name, ".."))
-            continue;
-
-        // Determinate a full path of an entry
-        char fullPath[256];
-        ASSERT(pathLength + strlen(entry->d_name) < ARRAY_COUNT(fullPath));
-        strcpy(fullPath, pathStr);
-        strcat(fullPath, "/");
-        strcat(fullPath, entry->d_name);
-
-        // Stat for the entry
-        stat(fullPath, &statEntry);
-
-        // Check for directory
-        if (S_ISDIR(statEntry.st_mode) != 0)
-        {
-            // Add directory
-            results.Add(String(fullPath));
-        }
-    }
-
-    closedir(dir);
-
-    return false;
-}
-
-bool AppleFileSystem::FileExists(const StringView& path)
-{
-    struct stat fileInfo;
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (stat(pathANSI.Get(), &fileInfo) != -1)
-    {
-        return S_ISREG(fileInfo.st_mode);
-    }
-    return false;
-}
-
-bool AppleFileSystem::DeleteFile(const StringView& path)
-{
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    return unlink(pathANSI.Get()) != 0;
-}
-
-uint64 AppleFileSystem::GetFileSize(const StringView& path)
-{
-    struct stat fileInfo;
-    fileInfo.st_size = 0;
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (stat(pathANSI.Get(), &fileInfo) != -1)
-    {
-        // Check for directories
-        if (S_ISDIR(fileInfo.st_mode))
-        {
-            fileInfo.st_size = 0;
-        }
-    }
-    return fileInfo.st_size;
-}
-
-bool AppleFileSystem::IsReadOnly(const StringView& path)
-{
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (access(pathANSI.Get(), W_OK) == -1)
-    {
-        return errno == EACCES;
-    }
-    return false;
-}
-
-bool AppleFileSystem::SetReadOnly(const StringView& path, bool isReadOnly)
-{
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    struct stat fileInfo;
-    if (stat(pathANSI.Get(), &fileInfo) != -1)
-    {
-        if (isReadOnly)
-        {
-            fileInfo.st_mode &= ~S_IWUSR;
-        }
-        else
-        {
-            fileInfo.st_mode |= S_IWUSR;
-        }
-        return chmod(pathANSI.Get(), fileInfo.st_mode) == 0;
-    }
-    return false;
-}
-
-bool AppleFileSystem::MoveFile(const StringView& dst, const StringView& src, bool overwrite)
-{
-    if (!overwrite && FileExists(dst))
-    {
-        // Already exists
-        return true;
-    }
-
-    if (overwrite)
-    {
-        unlink(StringAsANSI<>(*dst, dst.Length()).Get());
-    }
-    if (rename(StringAsANSI<>(*src, src.Length()).Get(), StringAsANSI<>(*dst, dst.Length()).Get()) != 0)
-    {
-        if (errno == EXDEV)
-        {
-            if (!CopyFile(dst, src))
-            {
-                unlink(StringAsANSI<>(*src, src.Length()).Get());
-                return false;
-            }
-        }
-        return true;
-    }
-    return false;
-}
 
 bool AppleFileSystem::CopyFile(const StringView& dst, const StringView& src)
 {
@@ -350,156 +74,6 @@ out_error:
         close(dstFile);
     errno = cachedError;
     return true;
-}
-
-bool AppleFileSystem::getFilesFromDirectoryTop(Array<String>& results, const char* path, const char* searchPattern)
-{
-    size_t pathLength;
-    struct stat statPath, statEntry;
-    struct dirent* entry;
-
-    // Stat for the path
-    stat(path, &statPath);
-
-    // If path does not exists or is not dir - exit with status -1
-    if (S_ISDIR(statPath.st_mode) == 0)
-    {
-        // Is not directory
-        return true;
-    }
-
-    // If not possible to read the directory for this user
-    DIR* dir = opendir(path);
-    if (dir == NULL)
-    {
-        // Cannot open directory
-        return true;
-    }
-
-    // The length of the path
-    pathLength = strlen(path);
-
-    // Iteration through entries in the directory
-    while ((entry = readdir(dir)) != NULL)
-    {
-        // Skip entries "." and ".."
-        if (!strcmp(entry->d_name, ".") || !strcmp(entry->d_name, ".."))
-            continue;
-
-        // Determinate a full path of an entry
-        char fullPath[256];
-        ASSERT(pathLength + strlen(entry->d_name) < ARRAY_COUNT(fullPath));
-        strcpy(fullPath, path);
-        strcat(fullPath, "/");
-        strcat(fullPath, entry->d_name);
-
-        // Stat for the entry
-        stat(fullPath, &statEntry);
-
-        // Check for file
-        if (S_ISREG(statEntry.st_mode) != 0)
-        {
-            // Validate with filter
-            const int32 fullPathLength = StringUtils::Length(fullPath);
-            const int32 searchPatternLength = StringUtils::Length(searchPattern);
-            if (searchPatternLength == 0 || StringUtils::Compare(searchPattern, "*") == 0)
-            {
-                // All files
-            }
-            else if (searchPattern[0] == '*' && searchPatternLength < fullPathLength && StringUtils::Compare(fullPath + fullPathLength - searchPatternLength + 1, searchPattern + 1, searchPatternLength - 1) == 0)
-            {
-                // Path ending
-            }
-            else
-            {
-                // TODO: implement all cases in a generic way
-                continue;
-            }
-
-            // Add file
-            results.Add(String(fullPath));
-        }
-    }
-
-    closedir(dir);
-
-    return false;
-}
-
-bool AppleFileSystem::getFilesFromDirectoryAll(Array<String>& results, const char* path, const char* searchPattern)
-{
-    // Find all files in this directory
-    getFilesFromDirectoryTop(results, path, searchPattern);
-
-    size_t pathLength;
-    DIR* dir;
-    struct stat statPath, statEntry;
-    struct dirent* entry;
-
-    // Stat for the path
-    stat(path, &statPath);
-
-    // If path does not exists or is not dir - exit with status -1
-    if (S_ISDIR(statPath.st_mode) == 0)
-    {
-        // Is not directory
-        return true;
-    }
-
-    // If not possible to read the directory for this user
-    if ((dir = opendir(path)) == NULL)
-    {
-        // Cannot open directory
-        return true;
-    }
-
-    // The length of the path
-    pathLength = strlen(path);
-
-    // Iteration through entries in the directory
-    while ((entry = readdir(dir)) != NULL)
-    {
-        // Skip entries "." and ".."
-        if (!strcmp(entry->d_name, ".") || !strcmp(entry->d_name, ".."))
-            continue;
-
-        // Determinate a full path of an entry
-        char full_path[256];
-        ASSERT(pathLength + strlen(entry->d_name) < ARRAY_COUNT(full_path));
-        strcpy(full_path, path);
-        strcat(full_path, "/");
-        strcat(full_path, entry->d_name);
-
-        // Stat for the entry
-        stat(full_path, &statEntry);
-
-        // Check for directory
-        if (S_ISDIR(statEntry.st_mode) != 0)
-        {
-            if (getFilesFromDirectoryAll(results, full_path, searchPattern))
-            {
-                closedir(dir);
-                return true;
-            }
-        }
-    }
-
-    closedir(dir);
-
-    return false;
-}
-
-DateTime AppleFileSystem::GetFileLastEditTime(const StringView& path)
-{
-    struct stat fileInfo;
-    const StringAsANSI<> pathANSI(*path, path.Length());
-    if (stat(pathANSI.Get(), &fileInfo) == -1)
-    {
-        return DateTime::MinValue();
-    }
-
-    const TimeSpan timeSinceEpoch(0, 0, 0, fileInfo.st_mtime);
-    return UnixEpoch + timeSinceEpoch;
 }
 
 void AppleFileSystem::GetSpecialFolderPath(const SpecialFolder type, String& result)
